Move refresh chat to helper file
